+++ packages/agentledger-sandbox-docker/src/agentledger_sandbox_docker/__init__.py
@@ -1,6 +1,6 @@
 from agentledger.sandbox import DockerSandboxExecutor, SandboxConfig, SandboxPolicy, SandboxResult, SandboxRouter, SandboxUnavailable
 
-__version__ = "1.4.2"
+__version__ = "1.5.0"
 
 __all__ = [
     "DockerSandboxExecutor",
@@ -11,4 +11,3 @@ __all__ = [
     "SandboxUnavailable",
     "__version__",
 ]
-